requirements.txt
fastapi==0.110.0
uvicorn[standard]
pydantic-settings
aiofiles
python-multipart
python-magic
websockets
httpx
cachetools
bech32
jinja2
